Almost fixed whammy
